split types.ts schemas file
