+++ frontend/src/ts/ape/adapters/ts-rest-adapter.ts
@@ -3,7 +3,7 @@ import { Method } from "axios";
 import { getIdToken } from "firebase/auth";
 import { envConfig } from "../../constants/env-config";
 import { getAuthenticatedUser, isAuthenticated } from "../../firebase";
-import type { Metadata } from "shared/schemas/types";
+import type { Metadata } from "shared/schemas/util";
 
 function buildApi(timeout: number): (args: ApiFetcherArgs) => Promise<{
   status: number;
