+++ shared/schemas/config.ts
@@ -1,5 +1,4 @@
 import { z } from "zod";
-import { ColorHexValueSchema } from "./types";
 import { token } from "./helpers";
 
 export const SmoothCaretSchema = z.enum(["off", "slow", "medium", "fast"]);
@@ -199,6 +198,23 @@ export type MinimumBurst = z.infer<typeof MinimumBurstSchema>;
 
 export const ShowAverageSchema = z.enum(["off", "speed", "acc", "both"]);
 export type ShowAverage = z.infer<typeof ShowAverageSchema>;
+
+export const ColorHexValueSchema = z.string().regex(/^#([\da-f]{3}){1,2}$/i);
+export type ColorHexValue = z.infer<typeof ColorHexValueSchema>;
+
+export const DifficultySchema = z.enum(["normal", "expert", "master"]);
+export type Difficulty = z.infer<typeof DifficultySchema>;
+
+export const NumberModeSchema = z.enum(["time", "words", "quote"]);
+export const CustomModeSchema = z.enum(["custom"]);
+export const ZenModeSchema = z.enum(["zen"]);
+
+export const ModeSchema = z.union([
+  NumberModeSchema,
+  CustomModeSchema,
+  ZenModeSchema,
+]);
+export type Mode = z.infer<typeof ModeSchema>;
 
 export const CustomThemeColorsSchema = z.tuple([
   ColorHexValueSchema,
