+++ shared/schemas/users.ts
@@ -1,0 +1,29 @@
+import { z } from "zod";
+import { DifficultySchema } from "./config";
+import { StringNumberSchema } from "./util";
+
+export const PersonalBestSchema = z.object({
+  acc: z.number().nonnegative().max(100),
+  consistency: z.number().nonnegative().max(100),
+  difficulty: DifficultySchema,
+  lazyMode: z.boolean().optional(),
+  language: z
+    .string()
+    .max(100)
+    .regex(/[\w+]+/),
+  punctuation: z.boolean().optional(),
+  numbers: z.boolean().optional(),
+  raw: z.number().nonnegative(),
+  wpm: z.number().nonnegative(),
+  timestamp: z.number().nonnegative(),
+});
+export type PersonalBest = z.infer<typeof PersonalBestSchema>;
+
+export const PersonalBestsSchema = z.object({
+  time: z.record(StringNumberSchema, z.array(PersonalBestSchema)),
+  words: z.record(StringNumberSchema, z.array(PersonalBestSchema)),
+  quote: z.record(StringNumberSchema, z.array(PersonalBestSchema)),
+  custom: z.record(z.literal("custom"), z.array(PersonalBestSchema)),
+  zen: z.record(z.literal("zen"), z.array(PersonalBestSchema)),
+});
+export type PersonalBests = z.infer<typeof PersonalBestsSchema>;
